Use uint16_t for unit name ID. Can remove win32_compat.h from ORACERES.h.

+++ src/ONEWSFRE.cpp
@@ -1044,7 +1044,7 @@ void News::general_die()
 
 
 	str  = "Votre général ";
-	str += race_res[short_para1]->get_single_name( (WORD)short_para2 );
+	str += race_res[short_para1]->get_single_name( static_cast<uint16_t>(short_para2) );
 	str += " est mort.";
 }
 //------- End of function News::general_die -----//
@@ -1078,12 +1078,12 @@ void News::unit_assassinated()
 	if( short_para3 == RANK_KING )
 	{
 		str  = "Votre Roi ";
-		str += race_res[short_para1]->get_single_name( (WORD)short_para2 );
+		str += race_res[short_para1]->get_single_name( static_cast<uint16_t>(short_para2) );
 	}
 	else
 	{
 		str  = "Votre général ";
-		str += race_res[short_para1]->get_single_name( (WORD)short_para2 );
+		str += race_res[short_para1]->get_single_name( static_cast<uint16_t>(short_para2) );
 	}
 
 	str += " a été assassiné par un espion ennemi.";
@@ -1215,7 +1215,7 @@ void News::lightning_damage()
 			str += "unités ";
 
 		if( short_para2 > 0 )
-			str += race_res[short_para2]->get_name((WORD) short_para3);
+			str += race_res[short_para2]->get_name(short_para3);
 		else
 			str += unit_res[short_para3]->name;
 		break;
@@ -1557,7 +1557,7 @@ char* News::nation_name1()
 	if( nation_name_id1 < 0 )		// human player - custom name
 		str += nation_array.get_human_name(nation_name_id1, 1);		// 1-first word of the name only
 	else
-		str += race_res[nation_race_id1]->get_single_name( (WORD)nation_name_id1 );
+		str += race_res[nation_race_id1]->get_single_name( static_cast<uint16_t>(nation_name_id1) );
 
 	//------ add color bar -------//
 
@@ -1582,7 +1582,7 @@ char* News::nation_name2()
 	if( nation_name_id2 < 0 )		// human player - custom name
 		str += nation_array.get_human_name(nation_name_id2, 1);		// 1-first word of the name only
 	else
-		str += race_res[nation_race_id2]->get_single_name( (WORD)nation_name_id2 );
+		str += race_res[nation_race_id2]->get_single_name( static_cast<uint16_t>(nation_name_id2) );
 
 	//------ add color bar -------//
 
@@ -1609,7 +1609,7 @@ char* News::king_name1(int addColor)
 	if( nation_name_id1 < 0 )		// human player - custom name
 		str = nation_array.get_human_name(nation_name_id1);
 	else
-		str = race_res[nation_race_id1]->get_name( (WORD)nation_name_id1 );
+		str = race_res[nation_race_id1]->get_name( static_cast<uint16_t>(nation_name_id1) );
 
 	//------ add color bar -------//
 
@@ -1638,7 +1638,7 @@ char* News::king_name2(int addColor)
 	if( nation_name_id2 < 0 )		// human player - custom name
 		str = nation_array.get_human_name(nation_name_id2);
 	else
-		str = race_res[nation_race_id2]->get_name( (WORD)nation_name_id2 );
+		str = race_res[nation_race_id2]->get_name( static_cast<uint16_t>(nation_name_id2) );
 
 	//------ add color bar -------//
