+++ src/ONEWSSPA.cpp
@@ -1042,7 +1042,7 @@ void News::general_die()
 	//----------------------------------------------//
 
 	str  = "Tu General, ";
-	str += race_res[short_para1]->get_single_name( (WORD)short_para2 );
+	str += race_res[short_para1]->get_single_name( static_cast<uint16_t>(short_para2) );
 	str += ", ha sido liquidado.";
 }
 //------- End of function News::general_die -----//
@@ -1076,12 +1076,12 @@ void News::unit_assassinated()
 	if( short_para3 == RANK_KING )
 	{
 		str  = "Tu Rey, ";
-		str += race_res[short_para1]->get_single_name( (WORD)short_para2 );
+		str += race_res[short_para1]->get_single_name( static_cast<uint16_t>(short_para2) );
 	}
 	else
 	{
 		str  = "Tu General, ";
-		str += race_res[short_para1]->get_single_name( (WORD)short_para2 );
+		str += race_res[short_para1]->get_single_name( static_cast<uint16_t>(short_para2) );
 	}
 
 	str += ", ha sido asesinado por un espía enemigo.";
@@ -1214,7 +1214,7 @@ void News::lightning_damage()
 			str += "unidad ";
 
 		if( short_para2 > 0 )
-			str += race_res[short_para2]->get_name((WORD) short_para3);
+			str += race_res[short_para2]->get_name(short_para3);
 		else
 			str += unit_res[short_para3]->name;
 		if( short_para5 )
@@ -1571,7 +1571,7 @@ char* News::nation_name1()
 	if( nation_name_id1 < 0 )		// human player - custom name
 		str += nation_array.get_human_name(nation_name_id1, 1);		// 1-first word of the name only
 	else
-		str += race_res[nation_race_id1]->get_single_name( (WORD)nation_name_id1 );
+		str += race_res[nation_race_id1]->get_single_name( static_cast<uint16_t>(nation_name_id1) );
 
 	//------ add color bar -------//
 
@@ -1596,7 +1596,7 @@ char* News::nation_name2()
 	if( nation_name_id2 < 0 )		// human player - custom name
 		str += nation_array.get_human_name(nation_name_id2, 1);		// 1-first word of the name only
 	else
-		str += race_res[nation_race_id2]->get_single_name( (WORD)nation_name_id2 );
+		str += race_res[nation_race_id2]->get_single_name( static_cast<uint16_t>(nation_name_id2) );
 
 	//------ add color bar -------//
 
@@ -1623,7 +1623,7 @@ char* News::king_name1(int addColor)
 	if( nation_name_id1 < 0 )		// human player - custom name
 		str = nation_array.get_human_name(nation_name_id1);
 	else
-		str = race_res[nation_race_id1]->get_name( (WORD)nation_name_id1 );
+		str = race_res[nation_race_id1]->get_name( static_cast<uint16_t>(nation_name_id1) );
 
 	//------ add color bar -------//
 
@@ -1652,7 +1652,7 @@ char* News::king_name2(int addColor)
 	if( nation_name_id2 < 0 )		// human player - custom name
 		str = nation_array.get_human_name(nation_name_id2);
 	else
-		str = race_res[nation_race_id2]->get_name( (WORD)nation_name_id2 );
+		str = race_res[nation_race_id2]->get_name( static_cast<uint16_t>(nation_name_id2) );
 
 	//------ add color bar -------//
